For the option 1: 1. Fix validation for end input date (for the option 1). 2. Fix the time-calculating bug. 3. Clean code for ReadInclusiveDateRange.java(deleted). 4. Add validation for ``Choose the type of Range``. 5. Arrange classes (Tabular, ChartDisplay)
