--- src/Covid19/Main.java
@@ -1,5 +1,0 @@
-package Covid19;
-
-public class Main {
-
-}
