+++ src/Covid19/UserInitialInput.java
@@ -19,7 +19,8 @@ public class UserInitialInput {
     String endInputDate;
     int groupingOption;
 
-    public UserInitialInput(String pathToNewCsv, String location, String chosenDate, String endInputDate, int dayAway, int groupingOption) {
+    public UserInitialInput(String pathToNewCsv, String location, String chosenDate, String endInputDate, int dayAway
+            , int groupingOption) {
         this.pathToNewCsv = pathToNewCsv;
         this.location = location;
         this.chosenDate = chosenDate;
@@ -40,7 +41,7 @@ public class UserInitialInput {
         String endInputDate = null;
         int dayAway = 0;
         if (DataChoice == 1) {
-            endInputDate = EndInputDateValidate();
+            endInputDate = EndInputDateValidate(pathToNewCSV);
         } else {
             dayAway = dayAwayValidate();
         }
@@ -60,27 +61,14 @@ public class UserInitialInput {
         }
     }
 
-    public static String EndInputDateValidate() {
-        Scanner input = new Scanner(System.in);
-        System.out.println("\n Please enter your desired end date in format MM/dd/yyyy: ");
-        while (true) {
-            try {
-                return input.nextLine();
-            } catch (Exception e) {
-                input.next();
-                System.out.println("Invalid Input! Please enter again: ");
-            }
-        }
-    }
 
     public static int dayAwayValidate() {
         Scanner input = new Scanner(System.in);
-        System.out.print("\n Enter the number of days that are away from the date you chose: ");
+        System.out.print("\n Enter the number of Days that are Away from the date you chose: ");
         while (true) {
             try {
                 return input.nextInt();
             } catch (Exception e) {
-                input.next();
                 System.out.print("Invalid Input! Please enter again: ");
             }
         }
@@ -96,18 +84,18 @@ public class UserInitialInput {
                 "\n (If it is not possible to divide groups equally, the program will raise an error). ");
         System.out.println();
         int groupingOption;
-        do {
-            System.out.print("\n Please enter only the number in those 3 options to choose: ");
-            while (true) {
-                try {
+
+        while (true) {
+            try {
+                do {
+                    System.out.print("Enter one of the options above: ");
                     groupingOption = input.nextInt();
-                    return groupingOption;
-                } catch (Exception e) {
-                    input.next();
-                    System.out.print("Invalid Input! Please enter again: ");
-                }
+                } while (groupingOption != 1 && groupingOption != 2 && groupingOption != 3);
+                return groupingOption;
+            } catch (Exception e) {
+                System.out.print("Invalid Input! Please enter again! ");
             }
-        } while (groupingOption != 1 && groupingOption != 2 && groupingOption != 3);
+        }
     }
 
     public static String locationValidate(String pathToNewCSV) throws IOException {
@@ -138,6 +126,55 @@ public class UserInitialInput {
         return location;
     }
 
+    public static String EndInputDateValidate(String pathToNewCSV) throws IOException, ParseException {
+        Scanner input = new Scanner(System.in);
+        String endDate = null;
+        String inputEndDate;
+        int count = 0;
+        SimpleDateFormat sdf = new SimpleDateFormat("M/d/yyyy");
+        Date smallestDate = sdf.parse("1/1/2020");
+        Date largestDate = sdf.parse("7/8/2021");
+
+        do {
+            BufferedReader csvReader = new BufferedReader(new FileReader(pathToNewCSV));
+            String row = "";
+            csvReader.readLine();
+            if (count >= 1) {
+                System.out.println("\n The date is out of our database time-range or not valid in your chosen " +
+                        "location. " +
+                        "Please type again!");
+            }
+            System.out.println("\n Please enter your desired END DATE in format MM/dd/yyyy: ");
+            Date chosenDateDateFormat = null;
+            //Parse Error: occurs when user input data that cannot be converted into date (ex: dsfnui, 234)
+            boolean error = true;
+            do {
+                System.out.print("The date should be in the range from 1/1/2020 to 7/8/2021: ");
+                inputEndDate = input.nextLine();
+                try {
+                    chosenDateDateFormat = sdf.parse(inputEndDate);
+                    error = false;
+                } catch (Exception e) {
+                    System.out.println("\n Please enter the date in this format 'M/d/yyyy' !!!");
+                }
+            } while (error);
+
+            if (endDate == null) {
+                while ((row = csvReader.readLine()) != null) {
+                    String[] data = row.split(",", -1);
+                    if (!smallestDate.after(chosenDateDateFormat) && !largestDate.before(chosenDateDateFormat)) {
+                        endDate = inputEndDate;
+                        break;
+                    }
+                }
+                count += 1;
+                csvReader.close(); //reset to the 1st line of csvFile
+            }
+        } while (endDate == null);
+        System.out.println("We've got '" + endDate + "' !!!");
+        return endDate;
+    }
+
     public static String chosenDateValidate(String pathToNewCSV) throws IOException, ParseException {
         Scanner input = new Scanner(System.in);
         String chosenDate = null;
@@ -152,10 +189,11 @@ public class UserInitialInput {
             String row = "";
             csvReader.readLine();
             if (count >= 1) {
-                System.out.println("\n The date is out of our database time-range or not valid in your chosen location. " +
+                System.out.println("\n The date is out of our database time-range or not valid in your chosen " +
+                        "location. " +
                         "Please type again!");
             }
-            System.out.println("\n Choose a date you want in this format 'M/d/yyyy'");
+            System.out.println("\n Please enter a START DATE you want in this format 'M/d/yyyy'");
             Date chosenDateDateFormat = null;
             //Parse Error: occurs when user input data that cannot be converted into date (ex: dsfnui, 234)
             boolean error = true;
